+++ src/posit/connect/users.py
@@ -8,10 +8,8 @@ from requests import Session
 from . import urls
 
 from .config import Config
+from .pagination import _MAX_PAGE_SIZE, PaginatedRequester
 from .resources import Resource
-
-# The maximum page size supported by the API.
-_MAX_PAGE_SIZE = 500
 
 
 class User(Resource, total=False):
@@ -57,36 +55,3 @@ class Users:
         url = urls.append_path(self.url, id)
         response = self.session.get(url)
         return User(**response.json())
-
-
-class PaginatedRequester:
-
-    def __init__(self, session, url, start_page=1, page_size=_MAX_PAGE_SIZE):
-        self.session = session
-        self.url = url
-        self.page_number = start_page
-        self.page_size = page_size
-        # The API response will tell us how many total entries there are,
-        # but we don't know yet.
-        self.total = None
-        self.seen = 0
-
-    def get_all(self) -> List[dict]:
-        # Do the first page, which will tell us how many results there are
-        result = self.get_next_page()
-        while self.seen < self.total:
-            self.page_number += 1
-            result += self.get_next_page()
-        return result
-
-    def get_next_page(self) -> List[dict]:
-        # Define query parameters for pagination.
-        params = {"page_number": self.page_number, "page_size": self.page_size}
-        # Send a GET request to the endpoint with the specified parameters.
-        response = self.session.get(self.url, params=params).json()
-        # On our first request, we won't have set the total yet, so do it
-        if self.total is None:
-            self.total = response["total"]
-        results = response["results"]
-        self.seen += len(results)
-        return results
